@@ -4,13 +4,14 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "jest",
-    "init": "nodemon ./dist/src -m prod",
-    "build": "tsc",
-    "db-menu-sql": "node ./scripts/menu.js --db sql",
-    "db-menu-mongo": "node ./scripts/menu.js --db mongo",
-    "db-init-sql": "node ./scripts/init.js",
-    "db-init-mongo": "node ./scripts/init.js"
+    "start": "npm install && npm run build && npm run prod",
+    "start:reprodgular": "npm install && npm run build && npm run prod",
+    "start:nodemon": "nodemon ./public/src -m dev",
+    "build": "tsc --project tsconfig.json",
+    "build:watch": "tsc --watch --project tsconfig.json",
+    "prod": "node ./public/src -m prod",
+    "dev": "concurrently \"npm run build:watch\" \"npm run start:nodemon\"",
+    "test": "jest"
   },
   "repository": {
     "type": "git",
@@ -22,39 +23,49 @@
     "url": "https://github.com/kode-neko/cheety-back/issues"
   },
   "homepage": "https://github.com/kode-neko/cheety-back#readme",
+  "devDependencies": {
+    "@faker-js/faker": "^7.5.0",
+    "@types/jest": "^29.1.0",
+    "@types/sinon": "^10.0.13",
+    "@types/supertest": "^2.0.12",
+    "@typescript-eslint/eslint-plugin": "^5.38.1",
+    "@typescript-eslint/parser": "^5.38.1",
+    "concurrently": "^7.4.0",
+    "cypress": "^10.9.0",
+    "eslint": "^8.24.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-config-airbnb-typescript": "^17.0.0",
+    "eslint-plugin-chai-expect": "^3.0.0",
+    "eslint-plugin-cypress": "^2.12.1",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-mocha": "^10.1.0",
+    "jest": "^29.1.1",
+    "nodemon": "^2.0.20",
+    "sinon": "^14.0.0",
+    "superagent": "^8.0.1",
+    "supertest": "^6.2.4",
+    "ts-jest": "^29.0.3",
+    "ts-node": "^10.9.1"
+  },
   "dependencies": {
+    "@eslint/create-config": "^0.3.1",
+    "@types/express": "^4.17.14",
     "@types/mongoose": "^5.11.97",
     "@types/node": "^18.7.23",
     "@types/yargs": "^17.0.13",
-    "bcrypt": "^5.0.1",
-    "chalk": "^4.1.2",
-    "cors": "^2.8.5",
-    "dotenv": "^16.0.2",
+    "bootstrap": "^5.2.1",
+    "bootstrap-icons": "^1.9.1",
+    "dotenv": "^16.0.3",
     "express": "^4.18.1",
-    "figlet": "^1.5.2",
-    "inquirer": "^8.2.4",
-    "mongodb": "^4.7.0",
-    "mongoose": "^6.6.1",
+    "i18next": "^21.9.2",
+    "i18next-http-middleware": "^3.2.1",
+    "i18next-node-fs-backend": "^2.1.3",
+    "mongo": "^0.1.0",
+    "mongodb": "^4.10.0",
+    "mongoose": "^6.6.2",
     "mysql": "^2.18.1",
-    "sequelize": "^6.23.1",
+    "sequelize": "^6.23.2",
+    "typescript": "^4.8.4",
     "yargs": "^17.5.1"
-  },
-  "devDependencies": {
-    "@faker-js/faker": "^7.5.0",
-    "@types/express": "^4.17.14",
-    "@types/jest": "^29.0.3",
-    "@types/sinon": "^10.0.13",
-    "@types/supertest": "^2.0.12",
-    "eslint": "^8.18.0",
-    "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-config-airbnb-typescript": "^17.0.0",
-    "eslint-plugin-import": "^2.26.0",
-    "jest": "^29.0.3",
-    "nodemon": "^2.0.20",
-    "sinon": "^14.0.0",
-    "superagent": "^8.0.0",
-    "supertest": "^6.2.4",
-    "ts-jest": "^29.0.1",
-    "typescript": "^4.7.4"
   }
 }
