@@ -1,11 +1,17 @@
 import { default as User } from './user';
-import { default as Article } from './article';
+import {
+  IArticle,
+  ArticleModel,
+  Article,
+} from './article';
 import { default as Tag } from './tag';
 import { default as ArticleLine } from './article-line';
 import { default as TagArticle } from './tag-article';
 
 export {
   User,
+  IArticle,
+  ArticleModel,
   Article,
   Tag,
   ArticleLine,
