@@ -2,22 +2,22 @@ import supertest from 'supertest';
 import { Response } from 'superagent';
 import app from '../../src/server';
 
-describe('Article Controler', () => {
-  it('Get all articles', async () => {
+describe('Article Ctrl', () => {
+  it('Ctrl GET /article', async () => {
     await supertest(app).get('/article')
       .expect(200)
       .then((res: Response) => {
         expect(res.body).toHaveLength(0);
       });
   });
-  it('Get article by id', async () => {
+  it('Ctrl GET /article/:id', async () => {
     await supertest(app).get(`/article/${1}`)
       .expect(200)
       .then((res: Response) => {
         expect(res.body).toEqual({});
       });
   });
-  it('Post article', async () => {
+  it('Ctrl POST /article', async () => {
     await supertest(app).post('/article')
       .send({})
       .expect(201)
@@ -25,7 +25,7 @@ describe('Article Controler', () => {
         expect(res.body).toEqual({});
       });
   });
-  it('Put article', async () => {
+  it('Ctrl PUT /article', async () => {
     await supertest(app).put('/article')
       .send({})
       .expect(200)
@@ -33,7 +33,7 @@ describe('Article Controler', () => {
         expect(res.body).toEqual({});
       });
   });
-  it('Delete article', async () => {
+  it('Ctrl DELETE /article/:id', async () => {
     await supertest(app).delete(`/article/${1}`)
       .expect(200)
       .then((res: Response) => {
